requirements.txt
Django==4.2
djangorestframework==3.14.0
cloudinary==1.29.0
boto3==1.24.35
botocore==1.27.35

